requirements.txt
fastapi
uvicorn
requests
python-dotenv  # Optional, for environment variables management
python-multipart
hhtpx